+++ crates/figma_import/src/lib.rs
@@ -27,7 +27,6 @@ mod fetch;
 mod figma_schema;
 mod image_context;
 pub mod reaction_schema;
-pub mod toolkit_layout_style;
 pub mod toolkit_schema;
 pub mod toolkit_style;
 pub mod tools;
