--- proto/definition/layout/display.proto
@@ -1,28 +1,0 @@
-// Copyright 2024 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-
-package designcompose.definition.layout;
-
-option java_multiple_files = true;
-option java_package = "com.android.designcompose.definition.layout";
-option optimize_for = LITE_RUNTIME;
-
-// Configures whether a view is displayed using a Flexbox layout or not at all.
-enum Display {
-  DISPLAY_UNSPECIFIED = 0;
-  DISPLAY_FLEX = 1;
-  DISPLAY_NONE = 2;
-}
