+++ proto/definition/view/view.proto
@@ -20,6 +20,7 @@ import "definition/element/geometry.proto";
 import "definition/element/view_shape.proto";
 import "definition/interaction/reaction.proto";
 import "definition/layout/grid.proto";
+import "definition/layout/positioning.proto";
 import "definition/plugin/frame_extras.proto";
 import "definition/view/component.proto";
 import "definition/view/text_style.proto";
